Create the new 'Context' class
This is the first step of my plan to collapse ContextPart and MethodContext.
Since BlockContext no longer exists, it makes no sense for them to be seperate
especially considering Cuis's goal of simplicity / minimalism.
The reason for not operating on the classes directly is that directly performing
surgery on such critical classes is very dangerous while this is safer.